backend/requirements.txt
Django==6.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
djangorestframework-simplejwt==5.5.1
PyJWT==2.12.1
